# spec/api.yml
openapi: 3.0.3
info:
  title: 診所預約系統 API
  description: |
    診所預約系統的 RESTful API 規格文件。

    每個 API 的 summary 欄位對應 Feature File 中的 step statement。
  version: 1.0.0
  contact:
    name: Clinic Booking System
servers:
  - url: /api
    description: API Base URL
...
paths:
...
  /admin/appointments/{appointmentId}:
    get:
      tags:
        - 管理員預約
      summary: 取得預約詳情
      operationId: getAdminAppointmentById
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 預約詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAppointmentResponse'

    put:
      tags:
        - 管理員預約
      summary: 管理員編輯預約
      description: |
        **Feature:** 管理員編輯預約.feature

        **Rules:**
        - 修改預約時必須立即釋放原時段分鐘數
        - 修改預約時必須檢查新時段餘量
        - 修改預約時必須扣除新時段分鐘數
        - 修改記錄必須包含操作人與操作時間
      operationId: updateAdminAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateAppointmentRequest'
      responses:
        '200':
          description: 預約記錄被更新，時段 {originalSlotId} 剩餘分鐘數為 {originalRemainingMinutes}，時段 {newSlotId} 剩餘分鐘數為 {newRemainingMinutes}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAppointmentResponse'
        '400':
          description: 操作失敗 - 新時段餘量不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # 管理員取消預約.feature
  # ============================================
    delete:
      tags:
        - 管理員預約
      summary: 管理員取消預約
      description: |
        **Feature:** 管理員取消預約.feature

        **Rules:**
        - 取消預約時必須更新預約狀態為「已取消」
        - 取消預約時必須釋放時段分鐘數
        - 取消成功後必須發送 LINE 通知
        - 可記錄取消原因
      operationId: cancelAdminAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentRequest'
      responses:
        '200':
          description: 預約狀態為 "cancelled"，時段 {slotId} 剩餘分鐘數為 {remainingMinutes}，LINE 訊息發送至 {lineUserId}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAppointmentResponse'

  # ============================================
  # 管理員更新預約狀態.feature
  # ============================================
  /admin/appointments/{appointmentId}/status:
    patch:
      tags:
        - 管理員預約
      summary: 管理員更新預約狀態為 {status}
      description: |
        **Feature:** 管理員更新預約狀態.feature

        **Rules:**
        - 已預約狀態可更新為已報到
        - 已報到狀態可更新為已完成
        - 已預約狀態不可直接更新為已完成
        - 已完成狀態不可變更
        - 已取消狀態不可變更
        - 未報到狀態可改為已報到（補報到）
        - 未報到無法改為已預約
        - 未報到無法改為已完成
      operationId: updateAppointmentStatus
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: 預約狀態為 {status}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAppointmentResponse'
        '400':
          description: 操作失敗 - 狀態轉換不允許
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # 管理病患資料.feature
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    AppointmentStatus:
      type: string
      enum: [booked, checked_in, completed, no_show, cancelled]
      description: |
        預約狀態：
        - booked: 已預約
        - checked_in: 已報到
        - completed: 已完成
        - no_show: 未報到
        - cancelled: 已取消
...
    AdminUpdateAppointmentRequest:
      type: object
      properties:
        slot_id:
          type: string
        treatment_id:
          type: string
    CancelAppointmentRequest:
      type: object
      properties:
        reason:
          type: string
          description: 取消原因
    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/AppointmentStatus'
    AdminAppointmentResponse:
      type: object
      properties:
        id:
          type: string
        patient_id:
          type: string
        patient_name:
          type: string
        patient_phone:
          type: string
        patient_line_user_id:
          type: string
        slot_id:
          type: string
        treatment_id:
          type: string
        treatment_name:
          type: string
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        status_label:
          type: string
        appointment_date:
          type: string
          format: date
        slot_start_time:
          type: string
        slot_end_time:
          type: string
        doctor_id:
          type: string
        doctor_name:
          type: string
        cancel_reason:
          type: string
        updated_by:
          type: string
        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    # ============================================
    # 管理員帳號
    # ============================================
...